backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-colorfield==0.8.0
djoser==2.1.0
drf-base64==2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
Pillow==9.3.0
PyJWT==2.1.0
pytz==2020.1
python-dotenv==0.21.0
sqlparse==0.3.1
weasyprint==57.2
